# docs/openapi/mobile_api.yaml
openapi: 3.1.0
info:
  title: Ratatoskr Mobile API
  description: RESTful API for Android/iOS mobile clients
  version: 1.0.0
servers:
  - url: https://ratatoskrapi.po4yka.com
    description: Production
  - url: https://staging-ratatoskrapi.po4yka.com
    description: Staging
  - url: http://localhost:8000
    description: Local development
paths:
...
  /v1/repositories:
    get:
      tags:
      - repositories
      summary: List Repositories
      description: List GitHub repositories for the authenticated user with optional filters and sort.
      operationId: list_repositories_v1_repositories_get
      security:
      - HTTPBearer: []
      parameters:
      - name: is_starred
        in: query
        required: false
        schema:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Starred
      - name: language
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            maxLength: 100
          - type: 'null'
          title: Language
      - name: topic
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            maxLength: 100
          - type: 'null'
          title: Topic
      - name: source
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            enum:
            - manual
            - starred
          - type: 'null'
          title: Source
      - name: pending_analysis
        in: query
        required: false
        schema:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Pending Analysis
      - name: sort
        in: query
        required: false
        schema:
          $ref: '#/components/schemas/RepositoryListSort'
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Repositories matching the filters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RepositoryListResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags:
      - repositories
      summary: Ingest Repository
      description: |
        Ingest a GitHub repository by URL. Validates the URL is a github.com
        repository URL, fetches metadata via the GitHub adapter, persists the
        Repository row, and (asynchronously) triggers analysis. Returns
        immediately with a repository id and a status of either `ready` (if
        the repository was processed inline) or `pending`.
      operationId: ingest_repository_v1_repositories_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/IngestRepositoryRequest'
      responses:
        '202':
          description: Ingestion accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/IngestRepositoryResponse'
        '400':
          $ref: '#/components/responses/BadRequestError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '502':
          $ref: '#/components/responses/InternalServerError'
  /v1/repositories/{repository_id}:
    get:
      tags:
      - repositories
      summary: Get Repository
      description: Return full detail for a single repository owned by the authenticated user.
      operationId: get_repository_v1_repositories_repository_id_get
      security:
      - HTTPBearer: []
      parameters:
      - name: repository_id
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: Repository detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RepositoryDetail'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'
    delete:
      tags:
      - repositories
      summary: Delete Repository
      description: Delete a repository and its Qdrant embedding point. Idempotent for missing rows (returns 404).
      operationId: delete_repository_v1_repositories_repository_id_delete
      security:
      - HTTPBearer: []
      parameters:
      - name: repository_id
        in: path
        required: true
        schema:
          type: integer
      responses:
        '204':
          description: Repository deleted (no content).
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/repositories/{repository_id}/reanalyze:
    post:
      tags:
      - repositories
      summary: Reanalyze Repository
      description: Force re-analysis of a repository. Re-runs the analysis agent and updates the stored analysis_json + analysis_at columns.
      operationId: reanalyze_repository_v1_repositories_repository_id_reanalyze_post
      security:
      - HTTPBearer: []
      parameters:
      - name: repository_id
        in: path
        required: true
        schema:
          type: integer
      responses:
        '200':
          description: Re-analysis completed and updated detail returned
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RepositoryDetail'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /v1/search/repositories:
    get:
      tags:
      - Search
      summary: Search Repositories
      description: Semantic search over GitHub repositories using Qdrant embeddings, filtered by optional language/topic/source/starred flags.
      operationId: search_repositories_v1_search_repositories_get
      security:
      - HTTPBearer: []
      parameters:
      - name: q
        in: query
        required: true
        schema:
          type: string
          minLength: 2
          maxLength: 200
          title: Q
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      - name: min_similarity
        in: query
        required: false
        schema:
          type: number
          maximum: 1.0
          minimum: 0.0
          default: 0.2
          title: Min Similarity
      - name: languages
        in: query
        required: false
        schema:
          anyOf:
          - type: array
            items:
              type: string
          - type: 'null'
          title: Languages
      - name: topics
        in: query
        required: false
        schema:
          anyOf:
          - type: array
            items:
              type: string
          - type: 'null'
          title: Topics
      - name: is_starred
        in: query
        required: false
        schema:
          anyOf:
          - type: boolean
          - type: 'null'
          title: Is Starred
      - name: source
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            enum:
            - manual
            - starred
          - type: 'null'
          title: Source
      responses:
        '200':
          description: Repository search results
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RepositorySearchResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '422':
          $ref: '#/components/responses/ValidationError'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer
  responses:
    BadRequestError:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    UnauthorizedError:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFoundError:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    ValidationError:
      description: Validation Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    InternalServerError:
      description: Internal Server Error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    Pagination:
      type: object
      description: Pagination metadata for list responses.
      required:
      - total
      - limit
      - offset
      - hasMore
      properties:
        total:
          type: integer
          description: Total number of items matching the query.
        limit:
          type: integer
          description: Maximum items per page.
        offset:
          type: integer
          description: Current offset from start.
        hasMore:
          type: boolean
          description: Whether more items are available.
        nextCursor:
          type: integer
          nullable: true
          description: Cursor for next page (if using cursor-based pagination).
...
    ErrorResponse:
      type: object
      required:
      - success
      - error
      - meta
      properties:
        success:
          type: boolean
          const: false
        error:
          $ref: '#/components/schemas/ErrorObject'
        meta:
          $ref: '#/components/schemas/Meta'
...
    IngestRepositoryRequest:
      type: object
      required:
      - url
      properties:
        url:
          type: string
          minLength: 10
          maxLength: 500
          description: github.com repository URL (any of the forms accepted by app.adapters.github.url_patterns.is_github_repo_url).
      title: IngestRepositoryRequest
      description: Request body for ingesting a GitHub repository by URL.
    RepositoryListSort:
      type: string
      enum:
      - stars_desc
      - pushed_desc
      - created_desc
      - full_name_asc
      title: RepositoryListSort
      description: Sort order for repository list.
    RepositoryAnalysis:
      type: object
      required:
      - purpose
      - tech_stack
      - architecture_summary
      - key_concepts
      - code_patterns
      - use_cases
      - target_audience
      - maturity
      - key_dependencies
      - hallucination_risk
      - confidence
      properties:
        purpose:
          type: string
        tech_stack:
          type: array
          items:
            type: string
        architecture_summary:
          type: string
        key_concepts:
          type: array
          description: List of "{term, explanation}" objects.
          items:
            type: object
            additionalProperties: true
        code_patterns:
          type: array
          description: List of "{name, description}" objects.
          items:
            type: object
            additionalProperties: true
        use_cases:
          type: array
          items:
            type: string
        target_audience:
          type: string
        maturity:
          type: string
        key_dependencies:
          type: array
          items:
            type: string
        hallucination_risk:
          type: string
        confidence:
          type: number
      title: RepositoryAnalysis
      description: Mirror of RepoAnalysis schema for API serialization.
    RepositoryCompact:
      type: object
      required:
      - id
      - github_id
      - full_name
      - owner
      - name
      - description
      - primary_language
      - stars
      - forks
      - is_starred
      - is_archived
      - pushed_at
      - last_synced_at
      - pending_analysis
      - has_analysis
      - source
      properties:
        id:
          type: integer
        github_id:
          type: integer
        full_name:
          type: string
        owner:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        primary_language:
          type: string
          nullable: true
        topics:
          type: array
          items:
            type: string
          default: []
        stars:
          type: integer
        forks:
          type: integer
        is_starred:
          type: boolean
        is_archived:
          type: boolean
        pushed_at:
          type: string
          format: date-time
          nullable: true
        last_synced_at:
          type: string
          format: date-time
        pending_analysis:
          type: boolean
        has_analysis:
          type: boolean
        source:
          type: string
          description: One of "manual" or "starred".
      title: RepositoryCompact
    RepositoryDetail:
      type: object
      required:
      - id
      - github_id
      - full_name
      - owner
      - name
      - description
      - primary_language
      - stars
      - forks
      - is_starred
      - is_archived
      - pushed_at
      - last_synced_at
      - pending_analysis
      - has_analysis
      - source
      - homepage_url
      - license_spdx
      - is_fork
      - is_template
      - readme_excerpt
      - analysis
      - analysis_model
      - analysis_at
      - content_hash
      - created_at_github
      - watchers
      properties:
        id:
          type: integer
        github_id:
          type: integer
        full_name:
          type: string
        owner:
          type: string
        name:
          type: string
        description:
          type: string
          nullable: true
        primary_language:
          type: string
          nullable: true
        topics:
          type: array
          items:
            type: string
          default: []
        stars:
          type: integer
        forks:
          type: integer
        is_starred:
          type: boolean
        is_archived:
          type: boolean
        pushed_at:
          type: string
          format: date-time
          nullable: true
        last_synced_at:
          type: string
          format: date-time
        pending_analysis:
          type: boolean
        has_analysis:
          type: boolean
        source:
          type: string
          description: One of "manual" or "starred".
        homepage_url:
          type: string
          nullable: true
        license_spdx:
          type: string
          nullable: true
        is_fork:
          type: boolean
        is_template:
          type: boolean
        languages:
          type: object
          additionalProperties:
            type: integer
          default: {}
        readme_excerpt:
          type: string
          nullable: true
        analysis:
          oneOf:
          - $ref: '#/components/schemas/RepositoryAnalysis'
          - type: 'null'
          nullable: true
        analysis_model:
          type: string
          nullable: true
        analysis_at:
          type: string
          format: date-time
          nullable: true
        content_hash:
          type: string
          nullable: true
        created_at_github:
          type: string
          format: date-time
          nullable: true
        watchers:
          type: integer
      title: RepositoryDetail
    RepositoryListResponse:
      type: object
      required:
      - repositories
      - pagination
      properties:
        repositories:
          type: array
          items:
            $ref: '#/components/schemas/RepositoryCompact'
        pagination:
          $ref: '#/components/schemas/Pagination'
      title: RepositoryListResponse
    IngestRepositoryResponse:
      type: object
      required:
      - repository_id
      - status
      - full_name
      properties:
        repository_id:
          type: integer
        status:
          type: string
          enum:
          - pending
          - ready
        full_name:
          type: string
      title: IngestRepositoryResponse
...
    RepositorySearchResponse:
      type: object
      required:
      - results
      - pagination
      - query
      properties:
        results:
          type: array
          items:
            $ref: '#/components/schemas/RepositorySearchHit'
        pagination:
          $ref: '#/components/schemas/Pagination'
        query:
          type: string
      title: RepositorySearchResponse

    # ------------------------------------------------------------------
    # GitHub auth (PAT + OAuth Device Flow)
    # ------------------------------------------------------------------
...